requirements.txt
pytest
selenium
chromedriver_autoinstaller
allure-pytest
openpyxl
pytest-ordering
pytest-xdist